Settings: Split code
Split code into their own views. This cleans up the overall project
and makes managing settings easier.
